rework concurrency (#1199)
