+++ ErsatzTV.Application/Emby/Commands/SynchronizeEmbyLibraries.cs
@@ -3,4 +3,4 @@
 namespace ErsatzTV.Application.Emby;
 
 public record SynchronizeEmbyLibraries(int EmbyMediaSourceId) : IRequest<Either<BaseError, Unit>>,
-    IEmbyBackgroundServiceRequest;
+    IScannerBackgroundServiceRequest;
