+++ ErsatzTV.Application/IScannerBackgroundServiceRequest.cs
@@ -1,0 +1,5 @@
+namespace ErsatzTV.Application;
+
+public interface IScannerBackgroundServiceRequest
+{
+}
