+++ ErsatzTV.Application/Jellyfin/Commands/SynchronizeJellyfinAdminUserId.cs
@@ -3,4 +3,4 @@
 namespace ErsatzTV.Application.Jellyfin;
 
 public record SynchronizeJellyfinAdminUserId(int JellyfinMediaSourceId) : IRequest<Either<BaseError, Unit>>,
-    IJellyfinBackgroundServiceRequest;
+    IScannerBackgroundServiceRequest;
