+++ ErsatzTV.Application/Jellyfin/Commands/SynchronizeJellyfinLibraries.cs
@@ -3,4 +3,4 @@
 namespace ErsatzTV.Application.Jellyfin;
 
 public record SynchronizeJellyfinLibraries(int JellyfinMediaSourceId) : IRequest<Either<BaseError, Unit>>,
-    IJellyfinBackgroundServiceRequest;
+    IScannerBackgroundServiceRequest;
