+++ ErsatzTV/Services/EmbyService.cs
@@ -4,8 +4,6 @@ using ErsatzTV.Application;
 using ErsatzTV.Application.Emby;
 using ErsatzTV.Core;
 using ErsatzTV.Core.Domain;
-using ErsatzTV.Core.Errors;
-using ErsatzTV.Core.Interfaces.Locking;
 using MediatR;
 
 namespace ErsatzTV.Services;
@@ -54,9 +52,6 @@ public class EmbyService : BackgroundService
                             break;
                         case SynchronizeEmbyLibraries synchronizeEmbyLibraries:
                             requestTask = SynchronizeLibraries(synchronizeEmbyLibraries, cancellationToken);
-                            break;
-                        case ISynchronizeEmbyLibraryById synchronizeEmbyLibraryById:
-                            requestTask = SynchronizeEmbyLibrary(synchronizeEmbyLibraryById, cancellationToken);
                             break;
                         default:
                             throw new NotSupportedException($"Unsupported request type: {request.GetType().Name}");
@@ -125,37 +120,5 @@ public class EmbyService : BackgroundService
                 "Unable to synchronize Emby libraries for source {MediaSourceId}: {Error}",
                 request.EmbyMediaSourceId,
                 error.Value));
-    }
-
-    private async Task SynchronizeEmbyLibrary(ISynchronizeEmbyLibraryById request, CancellationToken cancellationToken)
-    {
-        using IServiceScope scope = _serviceScopeFactory.CreateScope();
-        IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
-        IEntityLocker entityLocker = scope.ServiceProvider.GetRequiredService<IEntityLocker>();
-
-        Either<BaseError, string> result = await mediator.Send(request, cancellationToken);
-        result.BiIter(
-            name => _logger.LogDebug("Done synchronizing emby library {Name}", name),
-            error =>
-            {
-                if (error is ScanIsNotRequired)
-                {
-                    _logger.LogDebug(
-                        "Scan is not required for emby library {LibraryId} at this time",
-                        request.EmbyLibraryId);
-                }
-                else
-                {
-                    _logger.LogWarning(
-                        "Unable to synchronize emby library {LibraryId}: {Error}",
-                        request.EmbyLibraryId,
-                        error.Value);
-                }
-            });
-        
-        if (entityLocker.IsLibraryLocked(request.EmbyLibraryId))
-        {
-            entityLocker.UnlockLibrary(request.EmbyLibraryId);
-        }
     }
 }
