+++ ErsatzTV/Services/JellyfinService.cs
@@ -4,8 +4,6 @@ using ErsatzTV.Application;
 using ErsatzTV.Application.Jellyfin;
 using ErsatzTV.Core;
 using ErsatzTV.Core.Domain;
-using ErsatzTV.Core.Errors;
-using ErsatzTV.Core.Interfaces.Locking;
 using MediatR;
 
 namespace ErsatzTV.Services;
@@ -51,15 +49,6 @@ public class JellyfinService : BackgroundService
                     {
                         case SynchronizeJellyfinMediaSources synchronizeJellyfinMediaSources:
                             requestTask = SynchronizeSources(synchronizeJellyfinMediaSources, cancellationToken);
-                            break;
-                        case SynchronizeJellyfinAdminUserId synchronizeJellyfinAdminUserId:
-                            requestTask = SynchronizeAdminUserId(synchronizeJellyfinAdminUserId, cancellationToken);
-                            break;
-                        case SynchronizeJellyfinLibraries synchronizeJellyfinLibraries:
-                            requestTask = SynchronizeLibraries(synchronizeJellyfinLibraries, cancellationToken);
-                            break;
-                        case ISynchronizeJellyfinLibraryById synchronizeJellyfinLibraryById:
-                            requestTask = SynchronizeJellyfinLibrary(synchronizeJellyfinLibraryById, cancellationToken);
                             break;
                         default:
                             throw new NotSupportedException($"Unsupported request type: {request.GetType().Name}");
@@ -114,73 +103,5 @@ public class JellyfinService : BackgroundService
                     "Unable to synchronize jellyfin media sources: {Error}",
                     error.Value);
             });
-    }
-
-    private async Task SynchronizeLibraries(SynchronizeJellyfinLibraries request, CancellationToken cancellationToken)
-    {
-        using IServiceScope scope = _serviceScopeFactory.CreateScope();
-        IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
-
-        Either<BaseError, Unit> result = await mediator.Send(request, cancellationToken);
-        result.BiIter(
-            _ => _logger.LogInformation(
-                "Successfully synchronized Jellyfin libraries for source {MediaSourceId}",
-                request.JellyfinMediaSourceId),
-            error => _logger.LogWarning(
-                "Unable to synchronize Jellyfin libraries for source {MediaSourceId}: {Error}",
-                request.JellyfinMediaSourceId,
-                error.Value));
-    }
-
-    private async Task SynchronizeAdminUserId(
-        SynchronizeJellyfinAdminUserId request,
-        CancellationToken cancellationToken)
-    {
-        using IServiceScope scope = _serviceScopeFactory.CreateScope();
-        IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
-
-        Either<BaseError, Unit> result = await mediator.Send(request, cancellationToken);
-        result.BiIter(
-            _ => _logger.LogInformation(
-                "Successfully synchronized Jellyfin admin user id for source {MediaSourceId}",
-                request.JellyfinMediaSourceId),
-            error => _logger.LogWarning(
-                "Unable to synchronize Jellyfin admin user id for source {MediaSourceId}: {Error}",
-                request.JellyfinMediaSourceId,
-                error.Value));
-    }
-
-    private async Task SynchronizeJellyfinLibrary(
-        ISynchronizeJellyfinLibraryById request,
-        CancellationToken cancellationToken)
-    {
-        using IServiceScope scope = _serviceScopeFactory.CreateScope();
-        IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
-        IEntityLocker entityLocker = scope.ServiceProvider.GetRequiredService<IEntityLocker>();
-
-        Either<BaseError, string> result = await mediator.Send(request, cancellationToken);
-        result.BiIter(
-            name => _logger.LogDebug("Done synchronizing jellyfin library {Name}", name),
-            error =>
-            {
-                if (error is ScanIsNotRequired)
-                {
-                    _logger.LogDebug(
-                        "Scan is not required for jellyfin library {LibraryId} at this time",
-                        request.JellyfinLibraryId);
-                }
-                else
-                {
-                    _logger.LogWarning(
-                        "Unable to synchronize jellyfin library {LibraryId}: {Error}",
-                        request.JellyfinLibraryId,
-                        error.Value);
-                }
-            });
-
-        if (entityLocker.IsLibraryLocked(request.JellyfinLibraryId))
-        {
-            entityLocker.UnlockLibrary(request.JellyfinLibraryId);
-        }
     }
 }
