+++ ErsatzTV/Services/WorkerService.cs
@@ -3,13 +3,9 @@ using Bugsnag;
 using ErsatzTV.Application;
 using ErsatzTV.Application.Maintenance;
 using ErsatzTV.Application.MediaCollections;
-using ErsatzTV.Application.MediaSources;
 using ErsatzTV.Application.Playouts;
-using ErsatzTV.Application.Search;
 using ErsatzTV.Application.Subtitles;
 using ErsatzTV.Core;
-using ErsatzTV.Core.Errors;
-using ErsatzTV.Core.Interfaces.Locking;
 using MediatR;
 
 namespace ErsatzTV.Services;
@@ -48,7 +44,6 @@ public class WorkerService : BackgroundService
                 try
                 {
                     IMediator mediator = scope.ServiceProvider.GetRequiredService<IMediator>();
-                    IEntityLocker entityLocker = scope.ServiceProvider.GetRequiredService<IEntityLocker>();
 
                     switch (request)
                     {
@@ -62,40 +57,6 @@ public class WorkerService : BackgroundService
                                     "Unable to build playout {PlayoutId}: {Error}",
                                     buildPlayout.PlayoutId,
                                     error.Value));
-                            break;
-                        case IScanLocalLibrary scanLocalLibrary:
-                            Either<BaseError, string> scanResult = await mediator.Send(
-                                scanLocalLibrary,
-                                cancellationToken);
-
-                            scanResult.BiIter(
-                                name => _logger.LogDebug(
-                                    "Done scanning local library {Library}",
-                                    name),
-                                error =>
-                                {
-                                    if (error is ScanIsNotRequired)
-                                    {
-                                        _logger.LogDebug(
-                                            "Scan is not required for local library {LibraryId} at this time",
-                                            scanLocalLibrary.LibraryId);
-                                    }
-                                    else
-                                    {
-                                        _logger.LogWarning(
-                                            "Unable to scan local library {LibraryId}: {Error}",
-                                            scanLocalLibrary.LibraryId,
-                                            error.Value);
-                                    }
-                                });
-
-                            if (entityLocker.IsLibraryLocked(scanLocalLibrary.LibraryId))
-                            {
-                                entityLocker.UnlockLibrary(scanLocalLibrary.LibraryId);
-                            }
-                            break;
-                        case RebuildSearchIndex rebuildSearchIndex:
-                            await mediator.Send(rebuildSearchIndex, cancellationToken);
                             break;
                         case DeleteOrphanedArtwork deleteOrphanedArtwork:
                             _logger.LogInformation("Deleting orphaned artwork from the database");
